refactor(all): system 패키지 생성 및 기존 클래스 분리
