+++ src/main/java/lotto/domain/UserLotto.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
-import lotto.util.Constant;
-import lotto.util.IOMessage;
+import lotto.system.Constant;
+import lotto.system.IOMessage;
 import lotto.util.RandomNumGenerator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
