+++ src/main/java/lotto/system/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.system;
 
 public final class ExceptionMessage {
     private ExceptionMessage() {
