+++ src/main/java/lotto/system/IOMessage.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.system;
 
 public final class IOMessage {
     private IOMessage() {}
