+++ src/main/java/lotto/verifier/LottoVerifier.java
@@ -1,7 +1,7 @@
 package lotto.verifier;
 
-import lotto.util.Constant;
-import lotto.util.ExceptionMessage;
+import lotto.system.Constant;
+import lotto.system.ExceptionMessage;
 
 import java.math.BigInteger;
 import java.util.Arrays;
