+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.util.IOMessage;
+import lotto.system.IOMessage;
 import lotto.verifier.BonusNumberVerifier;
 import lotto.verifier.MoneyVerifier;
 import lotto.verifier.LottoVerifier;
