+++ src/test/java/lotto/verifier/LottoVerifierTest.java
@@ -1,6 +1,6 @@
 package lotto.verifier;
 
-import lotto.util.ExceptionMessage;
+import lotto.system.ExceptionMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
