+++ src/test/java/lotto/verifier/MoneyVerifierTest.java
@@ -3,7 +3,7 @@ package lotto.verifier;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 
-import lotto.util.ExceptionMessage;
+import lotto.system.ExceptionMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
